pyproject.toml
[tool.poetry]
name = "talk"
version = "1.0"
description = "Backend for talk.io"
authors = ["Anvesh Mishra <anveshmishra0508@gmail.com>"]
maintainers = [
    "Anvesh Mishra <anveshmishra0508@gmail.com>",
]
repository = "https://github.com/Anv3sh/tak.io"
readme = "README.md"
keywords = ["chat","voip"]
packages = [{ include = "talk", from = "src/backend" }]
include = ["src/backend/talk/*", "src/backend/talk/**/*"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.103.1"
typer = "^0.9.0"
gunicorn = "^21.2.0"
rich = "^13.4.2"
docstring-parser = "^0.15"
psycopg2-binary = "^2.9.6"
qdrant-client = "^1.3.0"
python-multipart = "^0.0.6"
sqlmodel = "^0.0.8"
orjson = "3.9.3"
multiprocess = "^0.70.14"
cachetools = "^5.3.1"
types-cachetools = "^5.3.0.5"
appdirs = "^1.4.4"
certifi = "^2023.5.7"
alembic = "^1.11.2"
python-jose = "^3.3.0"
jose = "^1.0.0"
types-python-jose = "^3.3.4.8"
passlib = "^1.7.4"
types-passlib = "^1.7.7.13"
email-validator = "^2.0.0.post2"
types-psycopg2 = "^2.9.21.11"
pydantic_sqlalchemy = "^0.0.9"

beautifulsoup4 = "^4.12.2"
pypdf = "^3.11.0"
pysrt = "^1.1.2"
fake-useragent = "^1.1.3"
pyarrow = "^12.0.0"
tiktoken = "~0.4.0"
wikipedia = "^1.4.0"
websockets = "^10.3"
psycopg = "^3.1.9"
psycopg-binary = "^3.1.9"
fastavro = "^1.8.0"
bcrypt = "^4.0.1 "
cryptography = "^41.0.3"
types-jmespath = "^1.0.2.7"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.2"
mypy = "^1.1.1"
ruff = "^0.0.254"
httpx = "*"
pytest = "^7.2.2"
types-requests = "^2.28.11"
requests = "^2.28.0"
pytest-cov = "^4.0.0"
types-appdirs = "^1.4.3.5"
types-pyyaml = "^6.0.12.8"
jinja2 = "^3.1.2 "


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = ["tests", "integration"]
console_output_style = "progress"
filterwarnings = ["ignore::DeprecationWarning"]
log_cli = true


[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
